refactor: break up context modules into files
